+++ keras_cv/models/backbones/mobilenet_v3/mobilenet_v3_backbone_test.py
@@ -21,7 +21,7 @@ from tensorflow import keras
 from keras_cv.models.backbones.mobilenet_v3.mobilenet_v3_backbone import (
     MobileNetV3Backbone,
 )
-from keras_cv.models.backbones.mobilenet_v3.mobilenet_v3_backbone import (
+from keras_cv.models.backbones.mobilenet_v3.mobilenet_v3_aliases import (
     MobileNetV3SmallBackbone,
 )
 from keras_cv.utils.train import get_feature_extractor
